requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-core==2.18.2
pydantic-settings==2.2.1
httpx==0.27.0
pytest==8.2.1
pytest-asyncio==0.23.6
factory-boy==3.3.0
httpx[cli]==0.27.0
psycopg2-binary==2.9.10
email_validator==2.2.0
typing-extensions>=4.6.1
dnspython>=2.0.0
idna>=2.0.0
python-multipart==0.0.20
aiosqlite==0.21.0